doc: bundling docs and a script for external apps
This work shows a developer how to bundle up an app
as a component that can be usabe in an SSG or web site

This commit is not perfected, but is considered good
enough to be continuing  development as we utilize
these new docs to try out and keep perfecting this
approach.

This commit also contains minor fixes to both the
README and the WHY docs, unrelated to bundling.

Changes to be committed:
	modified:   README.md
	renamed:    READ_WHY.md -> WHY.md
	modified:   apps/df-firebase-teaching-app/README.md
	new file:   apps/df-firebase-teaching-app/guides/BUNDLE_INTEGRATION.md
	new file:   apps/df-firebase-teaching-app/guides/example-integration.html
	new file:   guides/BUNDLE_DEPLOYMENT.md
	new file:   scripts/copy-app-bundle.sh